refactor: 💡 optimize rendering based on valueType
